requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
audioop-lts==0.2.2
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
bleach==6.3.0
build==1.3.0
cachetools==6.2.2
ccxt==4.5.22
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.3.5
click==8.3.1
cohere==5.20.0
coincurve==21.0.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
comtypes==1.4.13
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
duckduckgo_search==8.1.1
durationpy==0.10
edge-tts==7.2.3
executing==2.2.1
fastapi==0.121.3
fastavro==1.12.1
fastjsonschema==2.21.2
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
googlesearch-python==1.3.0
groq==0.33.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
kubernetes==34.1.0
lark==1.3.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
mmh3==5.2.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.0
narwhals==2.12.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.5.0
notebook_shim==0.2.4
numpy==2.3.4
oauthlib==3.3.1
onnxruntime==1.23.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.0
plotly==6.5.0
pluggy==1.6.0
posthog==5.4.0
primp==0.15.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pybase64==1.4.2
pycares==4.11.0
pycaw==20251023
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyparsing==3.2.5
pypdf==6.4.0
pyperclip==1.11.0
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.2.0
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6_sip==13.10.2
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
pytest==9.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.2
pywin32==311
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
screen-brightness-control==0.23.0
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SpeechRecognition==3.14.3
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.50.0
sympy==1.14.0
TA-Lib==0.6.8
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.9.0+cpu
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.20.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.38.0
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
widgetsnbextension==4.0.15
xgboost==3.1.2
yarl==1.22.0
zipp==3.23.0
